+++ CHANGELOG.md
@@ -9,6 +9,7 @@ v2.7.0 [unreleased]
 - [#580](https://github.com/latera/homs/pull/580) Add several upload file fields for single form.
 
 ### Refactoring
+- [#592](https://github.com/latera/homs/pull/592) Replace memcached with redis.
 - [#586](https://github.com/latera/homs/pull/586) Stub api container in tests with local api mock file and refactor tests.
 
 v2.6.12 [2021-08-20]