+++ CHANGELOG.md
@@ -8,6 +8,7 @@ v2.7.0 [unreleased]
 - [#571](https://github.com/latera/homs/pull/512) Add dynamic `disable_if` and `delete_if` options for select_table field.
 - [#580](https://github.com/latera/homs/pull/580) Add several upload file fields for single form.
 - [#587](https://github.com/latera/homs/pull/587) Add dynamic `disable_if` and `delete_if` options for datetime field.
+- [#593](https://github.com/latera/homs/pull/593) Add disable_if/delete_if for submit_select buttons.
 
 ### Refactoring
 - [#592](https://github.com/latera/homs/pull/592) Replace memcached with redis.