+++ CHANGELOG.md
@@ -15,6 +15,11 @@ v2.7.0 [unreleased]
 - [#586](https://github.com/latera/homs/pull/586) Stub api container in tests with local api mock file and refactor tests.
 - [#594](https://github.com/latera/homs/pull/594) Get rid of Webpacker & build all assets with Webpack 5.
 
+v2.6.13 [unreleased]
+-------------------
+### Bugfixes
+- [#595](https://github.com/latera/homs/pull/595) Fix substitution of empty values of BP variables in localized static fields.
+
 v2.6.12 [2021-08-20]
 -------------------
 ### Features